Extract idevicebackup2 into a library
