+++ src/devicebackup2.h
@@ -1,0 +1,6 @@
+#ifndef __DEVICEBACKUP2_H
+#define __DEVICEBACKUP2_H
+
+#include "libimobiledevice/devicebackup2.h"
+
+#endif
